backend/requirements.txt
Django==6.0.1
djangorestframework==3.16.1
django-cors-headers==4.9.0
djangorestframework-simplejwt==5.5.1
gunicorn==21.2.0
psycopg2-binary==2.9.11